+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This project won't have versionining per se. But it's format is based in the [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 2023-04-03
+
+### Added
+
+- Javascript utility to get all the links from an open trello checklist and print them out
+
+## 2023-04-02
+
+### Modified
+
+- Implemented concurrency for the CLI to get words from a pdf
+
 ## 2023-03-26
 
 ### Added